Fix boost::format when used with logs
Two problems have occurred in building this with the latest 3.10.
First, boost::format needs the #include <boost/format.hpp> header.
Second, there are problems when trying to use boost::format with the
version of format that is used with spdlog.  This resolves both issues
and also adds a required #include <bitset> that caused one file to no
longer compile.

Signed-off-by: Ed Beroset <beroset@ieee.org>